Revert "moved and renamed abstract class and changed fields to protected"
This reverts commit 99ae25ada22434ce7213b5009e1893727e37506e.
